requirements.txt
django==5.0.6
djangorestframework==3.15.2
pandas==2.2.2
psycopg2-binary==2.9.9
requests==2.32.3